feat: better popup-menu types
