+++ packages/popup-menu/PopupMenuItemStyles.ts
@@ -1,5 +1,8 @@
 import styled, { css } from 'styled-components'
-import { PopupMenuVariants } from './types'
+import { PopupMenuVariants } from './constants'
+import { InferStyledComponentPropsWithoutRef } from '@lidofinance/utils'
+
+export type PopupMenuItemStyleProps = InferStyledComponentPropsWithoutRef<typeof PopupMenuItemStyle>
 
 export const PopupMenuItemStyle = styled.button<{ $active: boolean }>`
   text-align: left;
