+++ packages/popup-menu/constants.ts
@@ -3,3 +3,9 @@ export const FOCUSABLE_ELEMENTS = [
   'button:not([disabled]):not([aria-hidden])',
   '[tabindex]:not([tabindex^="-"])',
 ]
+
+export enum PopupMenuVariant {
+  small,
+  default,
+}
+export type PopupMenuVariants = keyof typeof PopupMenuVariant
