--- packages/popup-menu/types.ts
@@ -1,23 +1,0 @@
-import { LidoComponentProps } from '@lidofinance/utils'
-import { PopoverProps } from '@lidofinance/popover'
-import React from 'react'
-export type { Theme } from '@lidofinance/theme'
-
-export enum PopupMenuVariant {
-  small,
-  default,
-}
-export type PopupMenuVariants = keyof typeof PopupMenuVariant
-
-export type PopupMenuProps = {
-  variant?: PopupMenuVariants
-} & PopoverProps
-
-export type PopupMenuItemProps = LidoComponentProps<
-  'button',
-  {
-    leftDecorator?: React.ReactNode
-    rightDecorator?: React.ReactNode
-    active?: boolean
-  }
->
